+++ infra/binder/src/main/java/org/apache/shardingsphere/infra/binder/engine/type/DDLStatementBindEngine.java
@@ -26,6 +26,7 @@ import org.apache.shardingsphere.infra.binder.engine.statement.ddl.CreateTableSt
 import org.apache.shardingsphere.infra.binder.engine.statement.ddl.CursorStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.ddl.DropTableStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.ddl.RenameTableStatementBinder;
+import org.apache.shardingsphere.infra.binder.engine.statement.ddl.TruncateStatementBinder;
 import org.apache.shardingsphere.infra.hint.HintValueContext;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.AlterIndexStatement;
@@ -36,6 +37,7 @@ import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.CursorS
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.DDLStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.DropTableStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.RenameTableStatement;
+import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.TruncateStatement;
 
 /**
  * DDL statement bind engine.
@@ -77,6 +79,9 @@ public final class DDLStatementBindEngine {
         }
         if (statement instanceof AlterIndexStatement) {
             return new AlterIndexStatementBinder().bind((AlterIndexStatement) statement, binderContext);
+        }
+        if (statement instanceof TruncateStatement) {
+            return new TruncateStatementBinder().bind((TruncateStatement) statement, binderContext);
         }
         return statement;
     }